Configure store with flowtype
