+++ lib/VariableInput/index.js
@@ -1,27 +1,15 @@
 'use babel';
 
 // @flow
-type Position = {
-  x: ?number,
-  y: ?number,
-};
 
-type State = {
-  position: Position,
-  isOpened: boolean,
-  varName: string,
-};
+import type { State, Action } from './types';
 
-type TestAction = { type: 'TEST' };
-type OpenAction = { type: 'VI_OPEN', position: Position, varName: string };
-type CloseAction = { type: 'VI_CLOSE' };
-type ChangeVarNameAction = { type: 'VI_CHANGE_VAR_NAME', newVarName: string };
-
+// InitialState
 export const initialState: State = { position: { x: null, y: null }, isOpened: false, varName: '' };
 
 export const reducer = (
   state: State = initialState,
-  action: TestAction | OpenAction | CloseAction | ChangeVarNameAction,
+  action: Action,
 ) => {
   switch (action.type) {
     case 'VI_OPEN':
