+++ lib/VariableInput/test.js
@@ -4,10 +4,6 @@
 import { reducer, initialState } from './index';
 
 describe('VariableInputReducer', () => {
-  test('initialState is Ok', () => {
-    expect(reducer(undefined, { type: 'TEST' })).toEqual(initialState);
-  });
-
   const closeState = {
     ...initialState,
   };
