+++ lib/factorx.js
@@ -5,15 +5,15 @@
 // eslint-disable-next-line max-len
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import { CompositeDisposable } from 'atom';
-import store from './store';
+// import store from './store';
 
 class FactorX {
   subscriptions: atom$CompositeDisposable;
   store: Object
 
-  constructor() {
-    this.store = store;
-  }
+  // constructor() {
+    // this.store = store;
+  // }
 
   activate() {
     this.subscriptions = new CompositeDisposable();
