+++ lib/store/store.js
@@ -1,0 +1,10 @@
+'use babel';
+
+// @flow
+import { createStore } from 'redux';
+import rootReducer from './state';
+import type { Store } from './state';
+
+const store: Store = createStore(rootReducer);
+
+export default store;
